### 2021-08-24 #3656 @danielskatz
After the paper builds, you could also use `@whedon check references` to check the references.  I'm not sure why it's not working now, but perhaps it's related to the pdf problem.
