### 2021-09-01 #3679 @whedon
[REVIEW]: Devicely: A Python package for reading, timeshifting and writing sensor data
**Submitting author:** @arianesasso (<a href="http://orcid.org/0000-0002-3669-4599">AM Sasso, A Sasso</a>)
**Repository:** <a href="https://github.com/hpi-dhc/devicely/" target ="_blank">https://github.com/hpi-dhc/devicely/</a>
**Version:** v1.1.0
