### 2021-09-01 #3679 @prashjha
Thank you @djmitche and @luciorq for agreeing to be reviewers. Please read the comments above in this thread and also reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Also, you can browse the closed "REVIEW" issues on the "joss-reviews" repository to get some ideas on how to complete the reviews. Thanks again and good luck!
