### 2021-09-04 #3679 @djmitche
@arianesasso this looks fantastic.  The contributing guide is particularly well-done.  Just one request: something that helps users figure out how to report a bug and what level of support they can expect.  This may just be a GitHub issue template, for example.
