### 2021-09-06 hpi-dhc/devicely#32 @arianesasso
Create a GitHub issue template for users
Request by @djmitche on https://github.com/openjournals/joss-reviews/issues/3679

"Just one request: something that helps users figure out how to report a bug and what level of support they can expect. This may just be a GitHub issue template, for example."
