### 2021-09-06 #3679 @arianesasso
@djmitche thank you for your feedback! I think that is a great idea, I create a GitHub issue in our repository to address that: https://github.com/hpi-dhc/devicely/issues/32 :)
