### 2021-09-07 hpi-dhc/devicely#37 @luciorq
[JOSS Review] @luciorq Meta Issue Tracker
Dear `devicely` authors,

I will be reviewing this repository as part of the [JOSS submission Review Process](https://github.com/openjournals/joss-reviews/issues/3679)  for [devicely](https://github.com/hpi-dhc/devicely), submitted by @arianesasso.
I will use the list below to aggregate and track individual Issues created in the [hpci-dhc/devicely Issue Tracker](https://github.com/hpi-dhc/devicely/issues) during the JOSS Review Process.

Issues:
- [x] #38
- [x] #39
- [x] #40
- [x] #41
- [x] #42
- [x] #43
- [x] #44
- [x] #45
- [x] #46
- [x] #47
- [x] #48
- [x] #49
- [x] #50

Remember that all issues are open to discussion and additional considerations are welcome!

I really hope to contribute to the general Scientific Software quality and reduce the technical debt associated with the maintenance of Open-Source Scientific projects.
