### 2021-09-08 #3679 @luciorq
Great Work for @arianesasso and fellow `devicely` team.

The software is already in a great state and tick most JOSS requirements out of the box.
The authors went the extra mile to enhance the contribution documentation and make the software usable by the broader community.

There are some minimal issues raised in the GitHub repository that I would like to be discussed.

As already linked up here, the issues created during Review are on the [`devicely` Issue Tracker](https://github.com/hpi-dhc/devicely/issues) and are being tracked by the list in this [Meta Issue](https://github.com/hpi-dhc/devicely/issues/37).

Relevant Link: <https://github.com/hpi-dhc/devicely/issues/37>
