### 2021-09-09 hpi-dhc/devicely#37 @arianesasso
Dear @luciorq, thank you for all your effort in reviewing our repository! You made really great comments and we will address each one of them. I am writing on the issues individually but I will let you know here when all of them have been addressed. Cheers!
