### 2021-09-25 hpi-dhc/devicely#32 @jostmorgenstern
I created two templates, one for bug reports and one for feature requests. They are very similar to the default templates.
Could you check if they are what you had in mind?
