### 2021-09-25 #3679 @arianesasso
We finished implementing the suggestions :).
We left the issues open for comments, could you please check them? Thank you!

@djmitche (hpi-dhc/devicely#32) and @luciorq (hpi-dhc/devicely#37)
