### 2021-09-25 hpi-dhc/devicely#37 @arianesasso
I believe we addressed all concerns, I would be grateful if you could take a look at the issues again :).
