### 2021-09-29 #3679 @luciorq
Most of my concerns have been discussed, addressed, and updated.
I will be updating/closing the issues during the next day, then will update here!
