### 2021-10-08 hpi-dhc/devicely#37 @luciorq
Congratulations on the effort and attention on the topics pointed above.

I am closing this issue as I don't have any additional considerations.

✅👋
