### 2021-10-08 #3679 @luciorq
Dear @prashjha & @arianesasso ,

I just reviewed all commits and closed the relevant issues.
I agree with @djmitche, there aren't any additional considerations. Also, the regenerated manuscript looks good to me.

Therefore, I highly recommend this work for publication.
