### 2021-10-31 #3679 @arfon
This should be updated now. The paper might take ~24 hours to show as updated on the JOSS site due to caching.
